Fix buggy gen_simple_symbol.
gen_symbol must guarantee that the first symbol character is not a minus
or plus if the next character COULD be a digit.  In the first list of
the oneof clause, gen_simple_symbol explicitly allows minus and plus as
the first character and follows that by a non-numeric list, which is ok,
but in the second list of the oneof clause, it specifies a list of
gen_non_numeric - which COULD be a minus or plus. Following the oneof
clause, the gen_symbol_char COULD be a digit, which produces an invalid
symbol.

Adding one character from a generator that explicitly excludes minus or
plus, and removing the non_empty restriction from the gen_symbol, seems
to fix the issue.  However, I am not sure how it affects the overall
testing, maybe by excluding some valid test cases.  Also, I don't have
QC so I used QC Mini, which may or may not be a fair substitution.